requirements.txt
fastapi==0.109.2
uvicorn[standard]
pydantic>=2.0.0,<3.0.0
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-huggingface==0.3.1
langchain-pinecone
langchain-pinecone
datasets
pinecone-client==3.2.2
torch
accelerate
bitsandbytes
python-dotenv
protobuf
blobfile
tiktoken
sentence-transformers>=2.6.0
transformers>=4.45.0
tokenizers>=0.19.1
sentence-transformers>=2.6.0
